Move concurrent map to shared
