+++ packages/envd/internal/process/manager.go
@@ -5,7 +5,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/e2b-dev/infra/packages/envd/internal/smap"
+	"github.com/e2b-dev/infra/packages/shared/pkg/smap"
 )
 
 type Manager struct {
