+++ packages/envd/internal/subscriber/manager.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gorilla/websocket"
 	"go.uber.org/zap"
 
-	"github.com/e2b-dev/infra/packages/envd/internal/smap"
+	"github.com/e2b-dev/infra/packages/shared/pkg/smap"
 )
 
 type Manager struct {
